🐞 "Page build failure": submodules use https
Don't use ssh to clone the submodules, use HTTPS.

fixes:
```
The page build failed for the `main` branch with the following error:

The submodule registered for `./themes/hugo-theme-bootstrap4-blog` could not be cloned. Make sure it's using https:// and that it's a public repo. For more information, see https://docs.github.com/github/working-with-github-pages/troubleshooting-jekyll-build-errors-for-github-pages-sites#invalid-submodule.
```