Concourse: add a section on locking down Concourse
The procedures are untested, but I'm hoping to test them in the near
future.

It was a gross oversight on my part to publish a blog post without
realizing that important keys were publicly available. Sigh. This
commit addresses that oversight.

Now that I'm comfortable with `helm upgrade`, I use that instead of
`helm delete` and `helm install`